backend/requirements.txt
Django
djangorestframework 
djangorestframework_simplejwt 
django-cors-headers
psycopg2-binary
python-dotenv
requests
